requirements.txt
django>=4.0.4
mysqlclient
python-dotenv
django-solo
pillow
django-tinymce==3.4.0
django-anymail[mailjet]

#blog
djangorestframework